fix(vue): source FormattingBar/ZoomControl export types from .ts modules
The Nuxt adapter typechecks with a plain tsc that resolves .vue imports
through an ambient default-export shim, so the public index.ts re-exporting
named type members (FormattingBarProps, ZoomControlProps, ZoomLevel) directly
from .vue SFCs failed with TS2614. Move these types (and the picker types they
reference: ParagraphAlignment, StyleOption) into colocated .ts modules, matching
the existing DocxEditor/types.ts convention, and re-export from there. Behaviour
and public type names/shapes are unchanged.
