+++ packages/vue/src/index.ts
@@ -18,12 +18,10 @@
 export { default as DocxEditor } from "./components/DocxEditor.vue";
 // Standalone chrome components (mirror React's index): a host can render these
 // outside DocxEditor. Both are controlled — they own no editor view.
-export { default as FormattingBar, type FormattingBarProps } from "./components/FormattingBar.vue";
-export {
-  default as ZoomControl,
-  type ZoomControlProps,
-  type ZoomLevel,
-} from "./components/ui/ZoomControl.vue";
+export { default as FormattingBar } from "./components/FormattingBar.vue";
+export type { FormattingBarProps } from "./components/FormattingBar.types";
+export { default as ZoomControl } from "./components/ui/ZoomControl.vue";
+export type { ZoomControlProps, ZoomLevel } from "./components/ui/ZoomControl.types";
 export {
   renderAsync,
   type DocxEditorHandle,
